[bugfix] 70%

+++ src/test/java/WowTable.java
@@ -1,5 +1,5 @@
 import net.yellowstrawberry.objectexpress.table.Table;
 
-public interface WowTable extends Table<Long, Sexy> {
+public interface WowTable extends Table<Integer, Sexy> {
 
 }
